add: main.dartを整理しました
