+++ lib/app.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:food_gram_app/gen/l10n/l10n.dart';
+import 'package:food_gram_app/router/router.dart';
+import 'package:food_gram_app/utils/text_form_borders.dart';
+
+class MyApp extends ConsumerWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return MaterialApp.router(
+      localizationsDelegates: const [
+        ...L10n.localizationsDelegates,
+      ],
+      supportedLocales: const [
+        ...L10n.supportedLocales,
+      ],
+      routerConfig: ref.watch(routerProvider),
+      title: 'Flutter Demo',
+      debugShowCheckedModeBanner: kReleaseMode,
+      theme: ThemeData(
+        inputDecorationTheme: const InputDecorationTheme(
+          contentPadding: EdgeInsets.all(15),
+          focusedBorder: TextFormBorders.textFormFocusedBorder,
+          enabledBorder: TextFormBorders.textFormEnabledBorder,
+          focusedErrorBorder: TextFormBorders.textFormErrorBorder,
+          errorBorder: TextFormBorders.textFormErrorBorder,
+        ),
+      ),
+    );
+  }
+}
